:construction: Work on player nicknames
Current method seems to fit nicely, and the custom name is saved on the entity out-of-the-box. But there's lots of places that still reference their GameProfile name. There is also the fact that their underlying username and UUID still exist on the client - so hiding YouTubers on servers wouldn't really work... :poop: